Split test cases and contained all in a separate folder.
