+++ test/test_ai.py
@@ -1,0 +1,2 @@
+# coding: utf-8
+
