+++ test/test_city.py
@@ -1,0 +1,2 @@
+# coding: utf-8
+
