+++ test/test_player.py
@@ -1,0 +1,2 @@
+# coding: utf-8
+
